% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_82z.R
\name{data_82z}
\alias{data_82z}
\title{to read the raw data}
\usage{
data_82z(filepath, filename)
}
\arguments{
\item{filepath}{name and path to the 82z archive}

\item{filename}{name of the file with raw data}
}
\description{
to read the raw data
}
