% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_name_82z.R
\name{data_name_82z}
\alias{data_name_82z}
\title{create colnames for data tibble}
\usage{
data_name_82z(filepath, data_file)
}
\arguments{
\item{filepath}{name and path to the 82z archive}

\item{data_file}{name of the file with raw data}
}
\value{
a character vector or the colnames of the data file
}
\description{
create colnames for data tibble
}
