% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_meta_82z.R
\name{df_meta_82z}
\alias{df_meta_82z}
\title{to extract the metadata that will be bound to the dataframe}
\usage{
df_meta_82z(metadata)
}
\description{
to extract the metadata that will be bound to the dataframe
}
