% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_id.R
\name{flux_id}
\alias{flux_id}
\title{gives a unique id to fluxes}
\usage{
flux_id(
  raw_conc,
  id_cols = c("Obs#", "File Name"),
  datetime_col = "Date",
  etime_col = "Etime",
  vol_tot_col,
  area_col,
  obs_len = "Observation Length"
)
}
\description{
gives a unique id to fluxes
}
