% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxible_81x.R
\name{fluxible_81x}
\alias{fluxible_81x}
\title{makes df from 81x files compatible with fluxible}
\usage{
fluxible_81x(df, focus_gas, id_cols, datetime_col)
}
\arguments{
\item{df}{input dataframe from \link[licoread:licoread]{licoread}}

\item{focus_gas}{gas to select}

\item{id_cols}{columns to identify unique fluxes}

\item{datetime_col}{column containing datetime information}
}
\value{
a df with the focus gas column renamed as "f_conc" and f_fluxid in
chronological order of datetime
}
\description{
makes df from 81x files compatible with fluxible
}
