% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxible_82z.R
\name{fluxible_82z}
\alias{fluxible_82z}
\title{makes df from 82z files compatible with fluxible}
\usage{
fluxible_82z(df, focus_gas)
}
\arguments{
\item{df}{input dataframe from \link[licoread:licoread]{licoread}}

\item{focus_gas}{gas to select}
}
\value{
an unnested df with only the selected gas
}
\description{
makes df from 82z files compatible with fluxible
}
