% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import7500.R
\name{import7500}
\alias{import7500}
\title{imports data from li7500 setup}
\usage{
import7500(path, version = "till2023", plotinfo = FALSE)
}
\arguments{
\item{path}{location of the files. Note that the function is recursive and
will import all the files at the location}

\item{version}{\code{"till2023"} (default) refers to the setup before the li7500
update. For the update version, use \code{"post2023"}.}

\item{plotinfo}{character vector of names to use when seperating the
plot info from the file name. Names in the file name have to be separated
with an underscore, ex: \verb{[location]_[date]_[time of day]_[trial].txt}.
\code{FALSE} (default) means no information will be fetched from the filename.}
}
\value{
a dataframe with all data present in the files to import. New
datetime and fluxid columns contain respectively the datetime of
measurements and their orginial filename. If comments are present, a column
will contain the comments. Measurement informations from the file
names are stored in new columns according to \code{plotinfo}.
}
\description{
imports data from the li7500 setup described in
Halbritter et al (2024)
}
\examples{
path_pftc7 <- system.file("extdata/pftc7", package = "licoread")
import7500(path_pftc7, version = "post2023")

path_pftc5 <- system.file("extdata/pftc5", package = "licoread")
import7500(path_pftc5,
plotinfo = c("site", "treatment", "date", "plot", "trial"))

path_co2fluxtent <- system.file("extdata/co2fluxtent",
package = "licoread")
import7500(path_co2fluxtent,
plotinfo = c("date", "location", "time_of_day", "trial"))
}
\references{
Halbritter, A.H., Vandvik, V., Cotner, S.H. et al.
Plant trait and vegetation data along a 1314 m elevation gradient with fire
history in Puna grasslands, Perú. Sci Data 11, 225 (2024).
https://doi.org/10.1038/s41597-024-02980-3
}
