% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import7500_old.R
\name{import7500_old}
\alias{import7500_old}
\title{imports old version of li7500 files}
\usage{
import7500_old(list_files)
}
\arguments{
\item{list_files}{list of files including path}
}
\value{
a df with the content of all the files and file name as id
}
\description{
imports old version of li7500 files
}
