% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import7500_old_oneobs.R
\name{import7500_old_oneobs}
\alias{import7500_old_oneobs}
\title{imports one file from li7500 (old)}
\usage{
import7500_old_oneobs(filepath, comment, skip)
}
\arguments{
\item{filepath}{path to the file}
}
\value{
a df with the content of the file
}
\description{
imports one file from li7500 (old)
}
