% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licoread_auto.R
\name{licoread_auto}
\alias{licoread_auto}
\title{finds out the file type for licoread}
\usage{
licoread_auto(file_list)
}
\arguments{
\item{file_list}{list of files found in the location}
}
\value{
a single character string indicating the file type
}
\description{
finds out the file type for licoread
}
