% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_82z.R
\name{metadata_82z}
\alias{metadata_82z}
\title{read meta data file inside 82z archive}
\usage{
metadata_82z(filepath, meta_file)
}
\arguments{
\item{filepath}{name and path to the 82z archive}

\item{meta_file}{name of the file with meta data}
}
\value{
a tibble with the metadata from one observation
}
\description{
read meta data file inside 82z archive
}
