% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_df.R
\name{names_df}
\alias{names_df}
\title{to get a vector of names of a df, matching a regex}
\usage{
names_df(df, regname = "name\\\\d")
}
\arguments{
\item{df}{the df to get the names from}

\item{regname}{the regex expression to match}
}
\value{
a df with the names of the meta df
}
\description{
to get a vector of names of a df, matching a regex
}
