% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneobs_81x.R
\name{oneobs_81x}
\alias{oneobs_81x}
\title{reading a single measurement from 81x file}
\usage{
oneobs_81x(start, end, all_obs, file)
}
\arguments{
\item{start}{line number at which the measurement starts}

\item{end}{line number at which the measurement ends}

\item{all_obs}{list of all the lines from the full file}

\item{file}{filepath to the 81x files}
}
\value{
a df with 1 row with the meta data of the measurement
and raw data nested
}
\description{
reads a single measurement from a licor .81x file
}
