% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_81x.R
\name{read_81x}
\alias{read_81x}
\title{reading several 81x files in a folder}
\usage{
read_81x(location)
}
\arguments{
\item{location}{location of the files}
}
\value{
a nested tibble with the meta data from each measurements as row and
the raw data nested
}
\description{
reads several .81x files with several measurements
}
\examples{
path <- system.file("extdata/81x", package = "licoread")
read_81x(path)
}
