% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_81x_onefile.R
\name{read_81x_onefile}
\alias{read_81x_onefile}
\title{reads 81x licor file}
\usage{
read_81x_onefile(file)
}
\arguments{
\item{file}{filepath the the 81x file to read}
}
\value{
a nested tibble with the meta data from each measurements as row and
the raw data nested
}
\description{
reads a .81x file with several measurements
}
