% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_82z.R
\name{units_82z}
\alias{units_82z}
\title{to create a nested tibble with the units of data}
\usage{
units_82z(filepath, data_file, data_name, filename)
}
\arguments{
\item{filepath}{name and path to the 82z archive}

\item{data_file}{name of the file with raw data}

\item{data_name}{vector of colnames}

\item{filename}{name of the 82z archive}
}
\value{
a tibble with the units of the variables contained in the raw data
}
\description{
to create a nested tibble with the units of data
}
