% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_82z.R
\name{data_82z}
\alias{data_82z}
\title{to read the raw data}
\usage{
data_82z(filepath, data_file, data_name, gases, filename)
}
\arguments{
\item{filepath}{name and path to the 82z archive}

\item{data_file}{name of the file with raw data}

\item{data_name}{vector of colnames}

\item{gases}{list of gases}

\item{filename}{name of the 82z archive}
}
\value{
a long df with the actual data contained in the data file
}
\description{
to read the raw data
}
