% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licoread.R
\name{licoread}
\alias{licoread}
\title{reads Li-COR files in a given location}
\usage{
licoread(
  location,
  file_type = "auto",
  file_type_list = c("82z", "81x"),
  data_file = "data.csv",
  meta_file = "metadata.json",
  regex_file = "(\\\\w*-)*\\\\w*(?=([.]82z$))"
)
}
\description{
reads Li-COR files in a given location
}
