% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneobs_82z.R
\name{oneobs_82z}
\alias{oneobs_82z}
\title{to read one measurement from the 82z archive}
\usage{
oneobs_82z(path, filename, data_file, meta_file)
}
\arguments{
\item{path}{path to the 82z archive}

\item{filename}{name of the 82z archive}

\item{data_file}{name of the file with raw data}

\item{meta_file}{name of the file with meta data}
}
\description{
to read one measurement from the 82z archive
}
