% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneobs_82z.R
\name{oneobs_82z}
\alias{oneobs_82z}
\title{to read one measurement from the 82z archive}
\usage{
oneobs_82z(filepath, filename = "data.csv", meta_file = "metadata.json")
}
\description{
to read one measurement from the 82z archive
}
