% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_82z.R
\name{read_82z}
\alias{read_82z}
\title{to read several 82z files}
\usage{
read_82z(path, data_file = "data.csv", meta_file = "metadata.json")
}
\arguments{
\item{path}{location of the 82z files}
}
\description{
to read several 82z files
}
